Updates location client demo
