+++ lost-demos/location-client-demo/pom.xml
@@ -25,6 +25,12 @@
             <version>4.1.1.4</version>
             <scope>provided</scope>
         </dependency>
+
+        <dependency>
+            <groupId>com.mapzen.android</groupId>
+            <artifactId>prefs-plus</artifactId>
+            <version>1.0.0-SNAPSHOT</version>
+        </dependency>
     </dependencies>
 
     <build>
