+++ lost-demos/location-client-demo/res/xml/preferences.xml
@@ -8,17 +8,17 @@
         android:summary="@string/mock_mode_summary"
         android:title="@string/mock_mode_title" />
 
-    <EditTextPreference
+    <com.mapzen.prefsplus.EditFloatPreference
         android:defaultValue="0.0"
         android:key="@string/mock_lat_key"
         android:title="@string/mock_lat_title" />
 
-    <EditTextPreference
+    <com.mapzen.prefsplus.EditFloatPreference
         android:defaultValue="0.0"
         android:key="@string/mock_lng_key"
         android:title="@string/mock_lng_title" />
 
-    <EditTextPreference
+    <com.mapzen.prefsplus.EditFloatPreference
         android:defaultValue="0.0"
         android:key="@string/mock_accuracy_key"
         android:title="@string/mock_accuracy_title" />
@@ -29,7 +29,7 @@
         android:summary="@string/mock_mode_gpx_summary"
         android:title="@string/mock_mode_gpx_title" />
 
-    <EditTextPreference
+    <com.mapzen.prefsplus.EditTextPlusPreference
         android:defaultValue="lost.gpx"
         android:key="@string/mock_gpx_file_key"
         android:title="@string/mock_gpx_file_title" />
