+++ lost-demos/location-client-demo/src/main/java/com/example/lost/locationclient/SettingsActivity.java
@@ -2,6 +2,7 @@ package com.example.lost.locationclient;
 
 import android.app.Activity;
 import android.os.Bundle;
+import android.preference.PreferenceFragment;
 import android.view.MenuItem;
 
 /**
@@ -24,5 +25,16 @@ public class SettingsActivity extends Activity {
         }
 
         return super.onOptionsItemSelected(item);
+    }
+
+    /**
+     * Settings Fragment
+     */
+    public static class SettingsFragment extends PreferenceFragment {
+        @Override
+        public void onCreate(Bundle savedInstanceState) {
+            super.onCreate(savedInstanceState);
+            addPreferencesFromResource(R.xml.preferences);
+        }
     }
 }
